Re-enable updater  - Patch tauri cli to workaround a tauri bug that prevented use from bundling frameworks in the updater  - Update tauri/api to 1.3 and tauri/cli 1.3.1  - Add target triple in the release artifacts name  - Use ubuntu-20.04 instead of ubuntu-latest  - Configure release workflow to build to all supported platforms  - Replace `tauri-apps/tauri-action` with simply calling `pnpm desktop build`  - Simplify tauri-patched-cli-js to only build the required macOS binaries  - Unify build.js and dev.js in a tauri.js build script, which now acts as an all around wrapper for the tauri/cli
