+++ .github/scripts/setup-system.sh
@@ -141,7 +141,7 @@ if [ "$SYSNAME" = "Linux" ]; then
     echo "Installing dependencies with apt..."
 
     # Tauri dependencies
-    DEBIAN_TAURI_DEPS="libwebkit2gtk-4.0-dev build-essential curl wget libssl-dev libgtk-3-dev libayatana-appindicator3-dev librsvg2-dev"
+    DEBIAN_TAURI_DEPS="libwebkit2gtk-4.0-dev build-essential curl wget libssl-dev libgtk-3-dev libayatana-appindicator3-dev librsvg2-dev patchelf"
 
     # FFmpeg dependencies
     DEBIAN_FFMPEG_DEPS="libavcodec-dev libavdevice-dev libavfilter-dev libavformat-dev libavutil-dev libswscale-dev libswresample-dev ffmpeg"
@@ -162,7 +162,7 @@ if [ "$SYSNAME" = "Linux" ]; then
     echo "Installing dependencies with pacman..."
 
     # Tauri deps https://tauri.studio/guides/getting-started/setup/linux#1-system-dependencies
-    ARCH_TAURI_DEPS="webkit2gtk base-devel curl wget openssl appmenu-gtk-module gtk3 libappindicator-gtk3 librsvg libvips"
+    ARCH_TAURI_DEPS="webkit2gtk base-devel curl wget openssl appmenu-gtk-module gtk3 libappindicator-gtk3 librsvg libvips patchelf"
 
     # Webkit2gtk requires gstreamer plugins for video playback to work
     ARCH_VIDEO_DEPS="gst-libav gst-plugins-base gst-plugins-good gst-plugins-bad gst-plugins-ugly"
@@ -176,8 +176,7 @@ if [ "$SYSNAME" = "Linux" ]; then
     # Protobuf compiler - https://github.com/archlinux/svntogit-packages/blob/packages/protobuf/trunk/PKGBUILD provides `libprotoc`
     ARCH_LIBP2P_DEPS="protobuf"
 
-    sudo pacman -Sy
-    sudo pacman -S --needed $ARCH_TAURI_DEPS $ARCH_FFMPEG_DEPS $ARCH_BINDGEN_DEPS $ARCH_LIBP2P_DEPS $ARCH_VIDEO_DEPS
+    sudo pacman -Sy --needed $ARCH_TAURI_DEPS $ARCH_FFMPEG_DEPS $ARCH_BINDGEN_DEPS $ARCH_LIBP2P_DEPS $ARCH_VIDEO_DEPS
   elif has dnf; then
     echo "Detected dnf!"
     echo "Installing dependencies with dnf..."
@@ -191,7 +190,7 @@ if [ "$SYSNAME" = "Linux" ]; then
     # Tauri dependencies
     # openssl is manually declared here as i don't think openssl and openssl-devel are actually dependant on eachother
     # openssl also has a habit of being missing from some of my fresh Fedora installs - i've had to install it at least twice
-    FEDORA_TAURI_DEPS="openssl-devel curl wget libappindicator-gtk3 librsvg2-devel"
+    FEDORA_TAURI_DEPS="openssl-devel curl wget libappindicator-gtk3 librsvg2-devel patchelf"
 
     # required for building the openssl-sys crate
     FEDORA_OPENSSL_SYS_DEPS="perl-FindBin perl-File-Compare perl-IPC-Cmd perl-File-Copy"
@@ -207,8 +206,6 @@ if [ "$SYSNAME" = "Linux" ]; then
 
     # Protobuf compiler
     FEDORA_LIBP2P_DEPS="protobuf-compiler"
-
-    sudo dnf update
 
     if ! sudo dnf install $FEDORA_37_TAURI_WEBKIT && ! sudo dnf install $FEDORA_36_TAURI_WEBKIT; then
       err 'We were unable to install the webkit2gtk4.0-devel/webkit2gtk3-devel package.' \
@@ -281,15 +278,21 @@ elif [ "$SYSNAME" = "Darwin" ]; then
     _success=$(gh_curl "${_gh_url}/${_sd_gh_path}/actions/workflows/ffmpeg.yml/runs?page=${_page}&per_page=100&status=success" \
       | jq -r '. as $raw | .workflow_runs | if length == 0 then error("Error: \($raw)") else .[] | .artifacts_url end' \
       | while IFS= read -r _artifacts_url; do
-        if _artificat_path="$(
+        if _artifact_path="$(
           gh_curl "$_artifacts_url" \
-            | jq --arg version "$FFMPEG_VERSION" --arg arch "$_arch" -r \
+            | jq --arg version "$FFMPEG_VERSION" --arg arch "$(
+              if [ "${TARGET:-}" = 'aarch64-apple-darwin' ]; then
+                echo 'arm64'
+              else
+                echo "$_arch"
+              fi
+            )" -r \
               '. as $raw | .artifacts | if length == 0 then error("Error: \($raw)") else .[] | select(.name == "ffmpeg-\($version)-\($arch)") | "suites/\(.workflow_run.id)/artifacts/\(.id)" end'
         )"; then
           # nightly.link is a workaround for the lack of a public GitHub API to download artifacts from a workflow run
           # https://github.com/actions/upload-artifact/issues/51
           # TODO: Use Github's private API when running on CI
-          if curl -LSs "https://nightly.link/${_sd_gh_path}/${_artificat_path}" | tar -xOf- | XZ_OPT='-T0' tar -xJf- -C "$_frameworks_dir"; then
+          if curl -LSs "https://nightly.link/${_sd_gh_path}/${_artifact_path}" | tar -xOf- | XZ_OPT='-T0' tar -xJf- -C "$_frameworks_dir"; then
             printf 'yes'
             exit
           fi
@@ -320,6 +323,21 @@ elif [ "$SYSNAME" = "Darwin" ]; then
     _header="${_header#"${_frameworks_dir}/FFMpeg.framework/Headers/"}"
     ln -s "../FFMpeg.framework/Headers/${_header}" "${_frameworks_dir}/include/${_header}"
   done
+
+  # Workaround while https://github.com/tauri-apps/tauri/pull/3934 is not merged
+  echo "Download patched tauri cli.js build..."
+  case "$_arch" in
+    x86_64)
+      _artifact_url="https://nightly.link/${_sd_gh_path}/actions/artifacts/684422923.zip"
+      ;;
+    arm64)
+      _artifact_url="https://nightly.link/${_sd_gh_path}/actions/artifacts/684422911.zip"
+      ;;
+    *)
+      err "Unsupported architecture: $_arch"
+      ;;
+  esac
+  curl -LSs "$_artifact_url" | tar -xf- -C "${_frameworks_dir}/bin"
 
   echo "Download protobuf build"
   _page=1
