+++ apps/desktop/package.json
@@ -6,10 +6,9 @@
 	"private": true,
 	"scripts": {
 		"vite": "vite",
-		"dev": "node ./src-tauri/scripts/dev.js",
+		"dev": "node ./src-tauri/scripts/tauri.js dev",
 		"dev-vite": "vite --clearScreen=false --mode development",
-		"tauri": "tauri",
-		"build": "node ./src-tauri/scripts/build.js",
+		"build": "node ./src-tauri/scripts/tauri.js build",
 		"dmg": "open ../../target/release/bundle/dmg/",
 		"typecheck": "tsc -b",
 		"lint": "eslint src --cache"
@@ -21,7 +20,7 @@
 		"@sd/interface": "workspace:*",
 		"@sd/ui": "workspace:*",
 		"@tanstack/react-query": "^4.24.4",
-		"@tauri-apps/api": "1.2.0",
+		"@tauri-apps/api": "1.3.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-router-dom": "6.9.0",
@@ -30,9 +29,8 @@
 	"devDependencies": {
 		"@iarna/toml": "^2.2.5",
 		"@sd/config": "workspace:*",
-		"@tauri-apps/cli": "1.2.3",
+		"@tauri-apps/cli": "1.3.1",
 		"@types/babel-core": "^6.25.7",
-		"@types/lodash.merge": "^4.6.7",
 		"@types/react": "^18.0.21",
 		"@types/react-dom": "^18.0.6",
 		"@vitejs/plugin-react": "^2.1.0",
