+++ apps/desktop/src-tauri/src/main.rs
@@ -83,7 +83,7 @@ async fn main() -> tauri::Result<()> {
 
 	let app = app
 		.setup(|app| {
-			// tauri::updater::builder(app.handle()).should_install(|_current, _latest| true);
+			tauri::updater::builder(app.handle()).should_install(|_current, _latest| true);
 
 			let app = app.handle();
